No need to rollback to older pipewire?
I was looking at https://pkgs.org/search/?q=pw-jack&on=files and it seems that the `pw-jack` binary was provided by `pipewire-libjack-0.3.13-4.fc33.x86_64.rpm` when Fedora 33 was born, but nowadays it comes from `pipewire-jack-audio-connection-kit-0.3.22-4.fc33.x86_64.rpm`.

It seems this one is a better solution for installing the dependency, because it doesn't require to downgrade pipewire. Actually this way I was able to install the dependency on Silverblue.